Fixed: Delete http method

+++ backEnd/src/controller/v1/administracion/representante.controller.js
@@ -62,11 +62,17 @@ export const updateRepresentante = async (req, res) => {
 
 export const deleteRepresentante = async (req, res) => {
   try {
-    const { id } = req.params
-    const data = await service.deleteRepresentante(id)
-    const id_auditoria = await deleteAuditoria({ entity: 'representante', user: req.user, id })
-    data.id_auditoria = id_auditoria
-    res.status(200).json(data)
+    const cedulas = await service.getCedulas()
+    if (!existRepresentante(cedulas, req.params.id)) {
+      return res.status(400).json({ message: 'Representante no existe' })
+    } else {
+      const { id } = req.params
+      const data = await service.deleteRepresentante(id)
+      console.log(data)
+      const id_auditoria = await deleteAuditoria({ entity: 'representante', user: req.user, body: data, id })
+      data.id_auditoria = id_auditoria
+      res.status(200).json(data)
+    }
   } catch (error) {
     res.status(500).json(error)
   }
